% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryadVersionCount.R
\name{dryad_version_count}
\alias{dryad_version_count}
\title{Count the Number of Dataset Versions on Dryad}
\usage{
dryad_version_count(doi)
}
\arguments{
\item{doi}{A character string specifying the dataset DOI, in the form of "https://doi.org/..."}
}
\value{
An integer indicating the number of versions of a dataset on dryad.
}
\description{
\code{dryad_version_count()} retrieves how many versions exist for a dataset on the Dryad data repository.
}
\details{
This function first queries the Dryad API to obtain all available versions of a dataset associated with the given DOI. The total number of dataset versions is extracted from the API response.
}
\examples{
\dontrun{
dryad_version_count("https://doi.org/10.5061/dryad.z08kprrk1")
}

}
\seealso{
\code{\link[=dryad_version_overview]{dryad_version_overview()}}
}
