% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryadVersionInormation.R
\name{dryad_version_overview}
\alias{dryad_version_overview}
\title{Get information about different Versions of a dataset.}
\usage{
dryad_version_overview(doi, ...)
}
\arguments{
\item{doi}{a character vector with one element in the form of "https://...".}

\item{...}{: title, authors(firstName, lastName, email, affiliation, affiliationROR, orcid, order, abstract),funders(organization, identifierType,identifier, awardNumber, awardDescription, order), keywords, fieldOfScience, methods, versionNumber, versionStatus, curationStatus, versionChanges, publicationDate, lastModificationDate, visibility, sharingLink, xxxxx_links.curies(name, href, templated), _links.self.href, _links.stash:dataset.href, _links.stash:files.href, _links.stash:download.href}
}
\value{
A nested dataframe.
}
\description{
dryad_version_overview() allows to view all information about a dataset, that is saved on dryad. The function allows to create a nested data.frame.
When selecting certain information beforehand, the download will be limited to those information.
}
\examples{
dryad_version_overview(https://doi.org/10.5061/dryad.z08kprrk1)
#Use ... to pass additional argumnents to dryad_version_overview().
}
\seealso{
\code{\link[=dryad_version_count]{dryad_version_count()}}
}
