% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryadFileDownloadID.R
\name{dryadFileDownloadID}
\alias{dryadFileDownloadID}
\title{Download a Specific File from Dryad based on its ID}
\usage{
dryadFileDownloadID(id, ext, path = "~/Downloads", name = "data")
}
\arguments{
\item{id}{A character string specifying the file ID.}

\item{ext}{A character string specifying the file extension.}

\item{path}{A character string specifying the location of the file. Default: "~/Downloads"}

\item{name}{A character string specifying the name of the file. Default: "data"}
}
\value{
Invisibly downloads a file, or \code{NULL} if no file was selected.
}
\description{
\code{dryadFileDownload()} allows downloading a file of a dataset hosted on Dryad directly by using its ID.
The user provides the file ID and the extension. The file will be downloaded into a prior determine folder under a prior determined name.
}
\details{
This function directly accesses the file based on its ID.
The user is prompted to download the file to their \verb{~/Downloads} folder if not specified otherwise.
}
\examples{
\dontrun{
dryadFileDownloadID("3833094", "xlsx")
}

}
\seealso{
\code{\link[=dryad_version_overview]{dryad_version_overview()}}, \code{\link[=dryadFileDownload]{dryadFileDownload()}}, \code{\link[=dryadFileReadID]{dryadFileReadID()}}
}
