% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryadFileRead.R
\name{dryadFileRead}
\alias{dryadFileRead}
\title{Read a file from dryad API into R}
\usage{
dryadFileRead(doi)
}
\arguments{
\item{doi}{a character vector with one element in the form of "https://...".}
}
\value{
Reads in the file or downloads it, if read-in is impossible.
}
\description{
dryadFileRead() allows to access a specific file from a specific version from a data set of dryad based on the doi for the datase.
}
\note{
the function is currently limited to ".csv" and ".xlsx" files. For every other file, there is an option do download it.
}
\examples{
dryadFileRead(https://doi.org/10.5061/dryad.z08kprrk1)
}
\seealso{
\code{\link[=dryad_version_overview]{dryad_version_overview()}}

\code{\link[=dryadFileReadID]{dryadFileReadID()}}
}
