% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryadFileReadID.R
\name{dryadFileReadID}
\alias{dryadFileReadID}
\title{Read a file from dryad API into R}
\usage{
dryadFileReadID(id, ext)
}
\arguments{
\item{id}{a character vector with one element in the form of "xxxxxx".}
}
\value{
Reads in the file.
}
\description{
dryadFileReadID() allows to access a specific file from dryad based on the id of that file.
}
\note{
the function is currently limited to ".csv" and ".xlsx" files.
}
\examples{
dryadFileReadID("3833094", "xslx")
}
\seealso{
\code{\link[=dryad_version_overview]{dryad_version_overview()}}

\code{\link[=dryadFileRead]{dryadFileRead()}}
}
