% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryadVersionCount.R
\name{dryad_version_count}
\alias{dryad_version_count}
\title{Get information on how many versions of a dataset exist.}
\usage{
dryad_version_count(doi)
}
\arguments{
\item{doi}{a character vector with one element in the form of "https://...".}
}
\value{
An integer.
}
\description{
dryad_version_count() easily retrieves the number of versions of a dataset in the dryad data repository.
}
\examples{
dryad_version_count(https://doi.org/10.5061/dryad.z08kprrk1)
}
\seealso{
\code{\link[=dryad_version_overview]{dryad_version_overview()}}
}
