% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryadVersionDownload.R
\name{dryadVersionDownload}
\alias{dryadVersionDownload}
\title{Download All Files from a Dryad Dataset Version}
\usage{
dryadVersionDownload(doi, path = "~/Downloads")
}
\arguments{
\item{doi}{A character string specifying the dataset DOI, in the form of "https://doi.org/..."}

\item{path}{A character string specifying the location of the file. Default: "~/Downloads"}
}
\value{
Invisibly downloads one or more files to the specified folder. Returns \code{NULL} if no version is selected.
}
\description{
\code{dryadVersionDownload()} downloads all files from a specific version of a Dryad dataset, given its DOI. The user selects the desired version if multiple versions are available. All files are saved to the specified directory.
}
\details{
This function interacts with the Dryad REST API to retrieve dataset version information and download all associated files for the selected version.
}
\examples{
\dontrun{
dryadVersionDownload("https://doi.org/10.5061/dryad.z08kprrk1")
}

}
\seealso{
\code{\link[=dryad_version_overview]{dryad_version_overview()}}, \code{\link[=dryadFileDownloadID]{dryadFileDownloadID()}}, \code{\link[=dryadFileDownload]{dryadFileDownload()}}
}
