% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryadDatasetSearch.R
\name{dryadDatasetSearch}
\alias{dryadDatasetSearch}
\title{Search dryad API for specific datasets}
\usage{
dryadDatasetSearch(
  p = NA,
  perp = 100,
  terms = NA,
  affiliation = NA,
  modifiedSince = NA,
  modifiedBefore = NA,
  tenant = NA,
  issn = NA,
  relatedWorkIdentifier = NA,
  relatedWorkRelationship = NA
)
}
\arguments{
\item{p}{an integer. Which page of results to view. Returns all pages, if not specified.}

\item{perp}{an integer. Number of results to return on each page. Defaults to 1000 (Maximum allowed).}

\item{terms}{a vector. A list of terms to be searched. If multiple terms are supplied, matches will only be returned for items that contain all terms. A term may be negated to indicate terms that should not be present in the results (e.g., cat -fish).}

\item{affiliation}{a character variable. ROR identifier specifying an institutional affiliation (https://ror.org) that must be present in the list of dataset authors. The identifier should be in the full "https" format}

\item{tenant}{a character variable. The abbreviation for a "tenant" organization in Dryad. This will automatically search all affiliations associated with the given tenant. If both a tenant and affiliation are specified, the tenant will be ignored.}

\item{issn}{a character variable. The journal ISS (https://portal.issn.org).}

\item{relatedWorkIdentifier}{a character variable. The identifier that is present in a related work.}

\item{relatedWorkRelationship}{a character variable. The type of relationship expressed by a related work.}

\item{modifiedSince:}{A timestamp for limiting results. Datasets will only be returned that have been modified since the given time. The time must be specified in ISO 8601 format, and the time zone must be set to UTC, e.g., 2020-10-08T10:24:53Z}

\item{modifiedBefore:}{A timestamp for limiting results. Datasets will only be returned that have been modified before the given time. The time must be specified in ISO 8601 format, and the time zone must be set to UTC, e.g., 2020-10-08T10:24:53Z.}
}
\value{
Dataframe of search results
}
\description{
dryadDatasetSearch() allows to access dryad API and search for one or multiple datasets based on prior specified criteria
}
\examples{
dryadDatasetSearch(terms=c("dog", "-cat"), p = 4)
}
\seealso{
\code{\link[=dryad_version_overview]{dryad_version_overview()}}

\code{\link[=dryadFileReadID]{dryadFileReadID()}}

\code{\link[=dryadFileRead]{dryadFileRead()}}

\code{\link[=dryadFileDownloadID]{dryadFileDownloadID()}}
}
