% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryadFileDownload.R
\name{dryadFileDownload}
\alias{dryadFileDownload}
\title{Download a file from dryad API into R}
\usage{
dryadFileDownload(doi, path = "~/Downloads", name = "data")
}
\arguments{
\item{doi}{a character vector with one element in the form of "https://...".}

\item{path}{a character vector with one element in the form of "....". Default: "~/Downloads"}

\item{name}{a character vector with one element in the form of ".....". Default: "data"}
}
\value{
Downloaded file.
}
\description{
dryadFileDownload() allows to access a specific file from dryad based on the doi of the dataset, choosing the file from the dataset.
}
\examples{
dryadFileDownload(https://doi.org/10.5061/dryad.z08kprrk1)
}
\seealso{
\code{\link[=dryad_version_overview]{dryad_version_overview()}}

\code{\link[=dryadFileReadID]{dryadFileReadID()}}

\code{\link[=dryadFileRead]{dryadFileRead()}}

\code{\link[=dryadFileDownloadID]{dryadFileDownloadID()}}
}
