% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryadFileDownloadID.R
\name{dryadFileDownloadID}
\alias{dryadFileDownloadID}
\title{Download a file from dryad API into R}
\usage{
dryadFileDownloadID(id, ext, path = "~/Downloads", name = "data")
}
\arguments{
\item{id}{a character vector with one element in the form of "xxxxxx".}

\item{ext}{a character vector with one element in the form of for example "md".}

\item{path}{a character vector with one element in the form of "....". Default: "~/Downloads"}

\item{name}{a character vector with one element in the form of ".....". Default: "data"}
}
\value{
Downloaded file.
}
\description{
dryadFileDownloadID() allows to access a specific file from dryad based on the ID of that file.
}
\examples{
dryadFileReadID("3833094", "xlsx")
}
\seealso{
\code{\link[=dryad_version_overview]{dryad_version_overview()}}

\code{\link[=dryadFileReadID]{dryadFileReadID()}}

\code{\link[=dryadFileRead]{dryadFileRead()}}

\code{\link[=dryadFileDownload]{dryadFileDownload()}}
}
