% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryadVersionDownload.R
\name{dryadVersionDownload}
\alias{dryadVersionDownload}
\title{Download all files of a version from dryad API into R}
\usage{
dryadVersionDownload(doi, path = "~/Downloads")
}
\arguments{
\item{doi}{a character vector with one element in the form of "https://...".}

\item{path}{a character vector with one element in the form of "....". Default: "~/Downloads"}
}
\value{
Downloaded file.
}
\description{
dryadVersionDownload() allows to access all files of a version from dryad based on the doi of the dataset.
}
\examples{
dryadVersionDownload(https://doi.org/10.5061/dryad.z08kprrk1)
}
\seealso{
\code{\link[=dryad_version_overview]{dryad_version_overview()}}

\code{\link[=dryadFileReadID]{dryadFileReadID()}}

\code{\link[=dryadFileRead]{dryadFileRead()}}

\code{\link[=dryadFileDownloadID]{dryadFileDownloadID()}}
}
